+++ HISTORY.rst
@@ -13,9 +13,14 @@ Release Notes
 PlatformIO Core 6
 -----------------
 
-**A professional collaborative platform for declarative, safety-critical, and test-driven embedded development.**
+Unlock the true potential of embedded software development with
+PlatformIO's collaborative ecosystem, embracing declarative principles,
+test-driven methodologies, and modern toolchains for unrivaled success.
 
-6.1.10 (2023-??-??)
+6.1.11 (2023-??-??)
+~~~~~~~~~~~~~~~~~~~
+
+6.1.10 (2023-08-11)
 ~~~~~~~~~~~~~~~~~~~
 
 * Resolved an issue that caused generated projects for `PlatformIO IDE for VSCode <https://docs.platformio.org/en/latest/integration/ide/vscode.html>`__ to break when the ``-iprefix`` compiler flag was used